Extract JPL class to it's own file in lib.
