+++ lib/__init__.py
@@ -1,10 +1,12 @@
 from .earth_data import EarthData
+from .jpl_data import JPLData
 from .source_folder import SourceFolder
 from .tile_file import TileFile
 from .tile_merger import TileMerger
 
 __all__ = [
     'EarthData',
+    'JPLData',
     'SourceFolder',
     'TileFile',
     'TileMerger'
